Se actualizan los controladores, modelos, config y el middleware.

+++ app/controllers/user.controller.js
@@ -1,25 +1,31 @@
 const User = require("../models/user.model");
 const Bootcamp = require("../models/bootcamp.model");
+const bcrypt = require('bcrypt');
 
-const createUser = async (user) => {
+
+const createUser = async (req, res) => {
+const {firstName, lastName, email, password} = req.body;
+const saltRounds = 10;
+const hashedPassword = await bcrypt.hash(password, saltRounds);
   try {
     const createdUser = await User.create({
-      firstName: user.firstName,
-      lastName: user.lastName,
-      email: user.email,
+      firstName,
+      lastName,
+      email,
+      password: hashedPassword
     });
-
-    console.log(`>> Se ha creado el usuario: ${JSON.stringify(createdUser, null, 4)}`);
-    return createdUser;
+    res.status(201).json({ message: "Usuario creado"});
   } catch (err) {
     console.log(`>> Error al crear el usuario ${err}`);
     throw err;
   }
 };
 
-const findUserById = async (userId) => {
+const findUserById = async (req, res) => {
+  const userId = req.params.id;
   try {
     const user = await User.findByPk(userId, {
+      attributes: { exclude: ["password", "createdAt","updatedAt"] }, 
       include: [
         {
           model: Bootcamp,
@@ -32,16 +38,17 @@ const findUserById = async (userId) => {
       ],
     });
 
-    return user;
+    res.status(200).json({ message: "Usuario encontrado", user });
   } catch (err) {
     console.log(`>> Error mientras se encontraba el usuario: ${err}`);
     throw err;
   }
 };
 
-const findAll = async () => {
+const findAll = async (req, res) => {
   try {
     const users = await User.findAll({
+      attributes: { exclude: ["password", "createdAt","updatedAt"] }, 
       include: [
         {
           model: Bootcamp,
@@ -54,19 +61,26 @@ const findAll = async () => {
       ],
     });
 
-    return users;
+   res.status(200).json({ message: "Usuarios encontrados", users});
   } catch (err) {
     console.log(">> Error Buscando los Usuarios: ", err);
     throw err;
   }
 };
 
-const updateUserById = async (userId, fName, lName) => {
+const updateUserById = async (req,res) => {
+  const userId = req.params.id;
   try {
+    const user = await User.findByPk(userId);
+
+    if (!user) {
+      return res.status(404).json({ message: "Usuario no encontrado" });
+    }
+  const {firstName, lastName} = req.body;
     const updatedUser = await User.update(
       {
-        firstName: fName,
-        lastName: lName,
+        firstName,
+        lastName
       },
       {
         where: {
@@ -75,29 +89,38 @@ const updateUserById = async (userId, fName, lName) => {
       }
     );
 
-    console.log(`>> Se ha actualizado el usuario: ${JSON.stringify(updatedUser, null, 4)}`);
-    return updatedUser;
+  res.status(200).json({ message: "Usuario actualizado"});
   } catch (err) {
     console.log(`>> Error mientras se actualizaba el usuario: ${err}`);
     throw err;
   }
 };
 
-const deleteUserById = async (userId) => {
+const deleteUserById = async (req, res) => {
+  const userId = req.params.id;
   try {
+    const user = await User.findByPk(userId);
+
+    if (!user) {
+      return res.status(404).json({ message: "Usuario no encontrado" });
+    }
     const deletedUser = await User.destroy({
       where: {
         id: userId,
       },
     });
 
-    console.log(`>> Se ha eliminado el usuario: ${JSON.stringify(deletedUser, null, 4)}`);
-    return deletedUser;
+    res.status(200).json({ message: "Usuario eliminado"});
   } catch (err) {
     console.log(`>> Error mientras se eliminaba el usuario: ${err}`);
     throw err;
   }
 };
+
+const signIn = async (req, res) => {
+    res.json({ code: 200, message: "Login correcto.", token: req.token });
+};
+
 
 module.exports = {
   createUser,
@@ -105,4 +128,5 @@ module.exports = {
   findAll,
   updateUserById,
   deleteUserById,
+  signIn,
 };
